Machine Learning/Road Segmentation with Unet/requirements.txt
imutils==0.5.4
matplotlib==3.5.2
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.3.0
torch==1.11.0+cu113
torchvision==0.12.0
tqdm==4.64.0
ttach==0.0.3
